Refactor file comparison logic to separate util file
