+++ scale_build/update_image.py
@@ -1,4 +1,3 @@
-import difflib
 import logging
 import os
 
@@ -11,33 +10,10 @@ from .image.manifest import update_file_path
 from .image.update import install_rootfs_packages, build_rootfs_image
 from .utils.logger import LoggingContext
 from .utils.paths import CHROOT_BASEDIR, LOG_DIR, REFERENCE_FILES, REFERENCE_FILES_DIR, RELEASE_DIR
+from .utils.reference_files import compare_reference_files
+
 
 logger = logging.getLogger(__name__)
-
-
-def compare_reference_files(cut_nonexistent_user_group_membership=False):
-    for reference_file in REFERENCE_FILES:
-        with open(os.path.join(REFERENCE_FILES_DIR, reference_file)) as f:
-            reference = f.readlines()
-
-        if cut_nonexistent_user_group_membership:
-            if reference_file == 'etc/group':
-                # `etc/group` on newly installed system can't have group membership information for users that have
-                # not been created yet.
-                with open(os.path.join(CHROOT_BASEDIR, 'etc/passwd')) as f:
-                    reference_users = {line.split(':')[0] for line in f.readlines()}
-
-                for i, line in enumerate(reference):
-                    bits = line.rstrip().split(':')
-                    bits[3] = ','.join([user for user in bits[3].split(',') if user in reference_users])
-                    reference[i] = ':'.join(bits) + '\n'
-
-        with open(os.path.join(CHROOT_BASEDIR, reference_file)) as f:
-            real = f.readlines()
-
-        diff = list(difflib.unified_diff(reference, real))
-
-        yield reference_file, diff[3:]
 
 
 def build_update_image():
